requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5
ultralytics>=8.0.0
Pillow>=9.0.0
